fix: changed CatalogueActivity to HomeActivity
