+++ app/src/main/AndroidManifest.xml
@@ -18,13 +18,7 @@
             android:label="@string/title_activity_view_my_book"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".activity.SignUpActivity" />
-        <activity android:name=".activity.CatalogueActivity" />
-<!--        <activity android:name=".activity.LoginActivity" />-->
-<!--        <activity android:name=".activity.CatalogueActivity" />-->
-<!--        <activity-->
-<!--            android:name=".activity.ViewMyBookActivity"-->
-<!--            android:label="@string/title_activity_view_my_book"-->
-<!--            android:theme="@style/AppTheme.NoActionBar" />-->
+        <activity android:name=".activity.HomeActivity" />
         <activity android:name=".activity.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
