+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/activity_main"
-    tools:context=".activity.MainActivity">
+    tools:context=".activity.HomeActivity">
 
     <FrameLayout
         android:id="@+id/container"
